Move trie to its own package
