+++ core/trie/trie_test.go
@@ -1,4 +1,4 @@
-package core
+package trie
 
 import (
 	"encoding/hex"
